Move ContextFlush tests
